[madoka-selectionhack] Madoka selection hack script
Request:

I select a line in the subtitle grid, then I activate the script which
selects that line to the end of the file and sets the first line as the
active line?

I can currently do this by selecting a line, hitting Shift+End, which
will add all lines to the end of the file, but it makes the last line
the active line. So then I need to scroll back up, hit Alt+L-Click on
the first line of the selection to re-activate it so I can
Alt+L-Click-Drag the whole selection in the audio window.

Time taken to write macro: 1:26, no testing whatsoever.