Typewriter: Correct logic for skipping override blocks
In the old logic, each character would be numbered by `n` in the
following manner:

char  ab{c}de
   n  1233334

`char` would be assigned to `active_char` and nowhere else for each
iteration where `n == i`. In this example, when `i == 3`, each character
of the override block is assigned to `active_char`, getting replaced
with each new iteration until the `d` is reached. Because of that, the
override block is lost because it didn't get saved anywhere else;
`start` only contains `ab` instead of the expected `ab{c}`.

When working with logic that can be expressed as a state machine, it is
often easier to write and reason about when iterating through states and
consuming the input as needed instead of iterating through the input and
keeping track of the state through flags. The code has been refactored
to follow this pattern.

By doing so, it becomes a relatively simple task to handle override
blocks at the start: when encountering a `{`, just consume input until
and including the next `}`, then proceed as usual.

As a special case, if there is a `{` and there are no `}` in the rest of
the string, treat the `{` as regular text; this is what renderers do.
The output of the script will not be valid, though, because once the
`{\alphaH&FF&}` shifts past the stray `{`, everything from the stray `{`
until the `}` of `{\alpha&HFF&}` will be considered an override block,
so the line will get progressively shorter. There is no good way to
handle this without the libass-specific `\{` tag.

An alternative approach to writing a state machine is to tokenize the
input into text and override segments before the outermost loop and
construct `start`, `active_char` and `rest` from these segments. This
approach would be more efficient, but it is also more complex and the
performance gain is likely negligible.