Typewriter: Treat text tags as single characters
Remove the hackfix for non-override tags by treating them as single
characters. That means that `char`/`active_char` can no longer be
assumed to be a single Unicode character, but it also means that it's
unnecessary to peek at the next character after tokenization.

Perhaps more importantly, treating text tags as single characters means
that only one line is generated per text tag instead of two.

With this approach, it should now be trivial to support the libass text
tag `\{`.